Just some docs and cleanup
