+++ kaylee/Cargo.toml
@@ -3,8 +3,6 @@ name = "kaylee"
 version = "0.1.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [dependencies]
-#strum = "0.24.0"
 anyhow = "1.0.56"
 kaylee_derive = { path = "../kaylee_derive" }
