+++ kaylee/src/instructions/library.rs
@@ -1,0 +1,3 @@
+//! Instructions for the "standard library"
+//! @todo: This and system may be merged, there is a lot of overlap in my mind
+//! Opcodes reserved: 180 - 219
