+++ kaylee/src/instructions/logical.rs
@@ -1,0 +1,2 @@
+//! Instructions for performing logical operations (shift, or, and, etc)
+//! Opcodes reserved: 120 - 129
