+++ kaylee/src/instructions/misc.rs
@@ -1,0 +1,2 @@
+//! Instructions for Misc or overflow operations
+//! Opcodes reserved: 220 - 255
