+++ kaylee/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod compiler;
 pub mod shared;
 pub mod vm;
 pub mod repl;
