+++ kaylee_derive/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "kaylee_derive"
-version = "0.0.0"
+version = "0.1.0"
 authors = ["Michael Wilson <mwilson@electricjones.me>"]
 description = "Macross for Kaylee"
 edition = "2021"
